+++ src/LanguageServer/Protocol/Handler/Diagnostics/DiagnosticsRefreshQueue.cs
@@ -19,14 +19,14 @@ internal sealed class DiagnosticsRefreshQueue : AbstractRefreshQueue
     {
         private readonly IAsynchronousOperationListenerProvider _asyncListenerProvider;
         private readonly LspWorkspaceRegistrationService _lspWorkspaceRegistrationService;
-        private readonly Refresher _refresher;
+        private readonly IDiagnosticsRefresher _refresher;
 
         [ImportingConstructor]
         [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
         public Factory(
             IAsynchronousOperationListenerProvider asynchronousOperationListenerProvider,
             LspWorkspaceRegistrationService lspWorkspaceRegistrationService,
-            Refresher refresher)
+            IDiagnosticsRefresher refresher)
         {
             _asyncListenerProvider = asynchronousOperationListenerProvider;
             _lspWorkspaceRegistrationService = lspWorkspaceRegistrationService;
@@ -42,44 +42,14 @@ internal sealed class DiagnosticsRefreshQueue : AbstractRefreshQueue
         }
     }
 
-    [Shared]
-    [Export(typeof(Refresher))]
-    [Export(typeof(IDiagnosticsRefresher))]
-    internal sealed class Refresher : IDiagnosticsRefresher
-    {
-        /// <summary>
-        /// Incremented every time a refresh is requested.
-        /// </summary>
-        private int _globalStateVersion;
-
-        public event Action? WorkspaceRefreshRequested;
-
-        [ImportingConstructor]
-        [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
-        public Refresher()
-        {
-        }
-
-        public void RequestWorkspaceRefresh()
-        {
-            // bump version before sending the request to the client:
-            Interlocked.Increment(ref _globalStateVersion);
-
-            WorkspaceRefreshRequested?.Invoke();
-        }
-
-        public int GlobalStateVersion
-            => _globalStateVersion;
-    }
-
-    private readonly Refresher _refresher;
+    private readonly IDiagnosticsRefresher _refresher;
 
     private DiagnosticsRefreshQueue(
         IAsynchronousOperationListenerProvider asynchronousOperationListenerProvider,
         LspWorkspaceRegistrationService lspWorkspaceRegistrationService,
         LspWorkspaceManager lspWorkspaceManager,
         IClientLanguageServerManager notificationManager,
-        Refresher refresher)
+        IDiagnosticsRefresher refresher)
         : base(asynchronousOperationListenerProvider, lspWorkspaceRegistrationService, lspWorkspaceManager, notificationManager)
     {
         _refresher = refresher;
